Be more careful about the temporary file variable here. Test that we did anything to it before testing existence on disk
svn path=/src/trunk/Pootle/; revision=12569

+++ local_apps/pootle_app/models/translation_project.py
@@ -298,6 +298,7 @@ class TranslationProject(models.Model):
 
     def get_archive(self, stores):
         """returns an archive of the given filenames"""
+        tempzipfile = None
         try:
             # using zip command line is fast
             from tempfile import mkstemp
@@ -309,7 +310,7 @@ class TranslationProject(models.Model):
             if filedata:
                 return filedata
         finally:
-            if os.path.exists(tempzipfile):
+            if tempzipfile is not None and os.path.exists(tempzipfile):
                 os.remove(tempzipfile)
 
         # but if it doesn't work, we can do it from python
